Run GUI and REST APIs from shell + Workflow cleanup and logging added

+++ otto/ryu/network_state_db/network_db_operator.py
@@ -1,19 +1,22 @@
 from typing import Union
-
-from pymongo import MongoClient
+from bson import ObjectId
+from pymongo import MongoClient, InsertOne
 from pymongo.errors import PyMongoError
 
 from exceptions import (NetworkDatabaseException,
                         SwitchDocumentNotFound)
+from otto.ryu.network_state_db.network_state_finder import NetworkStateFinder
 
 
 class NetworkDbOperator:
     _MongoConnector: Union[None, MongoClient]
+    _network_state_finder: NetworkStateFinder
 
     def __init__(self):
         self._MongoConnector = None
         self._network_state_db = None
         self._switch_collection = None
+        self._network_state_finder = NetworkStateFinder()
 
     def connect(self):
         """
@@ -27,6 +30,36 @@ class NetworkDbOperator:
             self._network_state_db = self._MongoConnector["topology"]
             self._switch_collection = self._network_state_db["switches"]
 
+    def get_network_state(self) -> dict:
+        """
+        Method to find the current network state. Uses NetworkStateFinder to get a list of current
+        switches found in the network. Once the switches have been found, use the get_switch_details method
+        to create a document to be inputted into the topology.switches collection.
+
+        Yields:
+            dict: document for a switch to be inputted into topology.switches
+        """
+        found_switches = self._network_state_finder.get_switches()
+
+        for switch in found_switches:
+            switch_info = self._network_state_finder.get_switch_details(str(switch))
+            yield switch_info
+
+    def create_network_state_db(self):
+        """
+        Method to create the switches collection in otto_network_state_db.topology. This method should be called
+        by any child class of ControllerEnvironment when creating the initial network_state_db to be used for the remainder
+        of the program lifetime. The topology.switches collection will be updated periodically by the NetworkStateUpdater
+        """
+        self.drop_database()  # delete topology collection if it already exists
+
+        switches  = []
+
+        for found_switch in self.get_network_state():
+            found_switch['_id'] = ObjectId()
+            switches.append(InsertOne(found_switch))
+
+        self._switch_collection.bulk_write(switches)
     def put_switch_to_db(self, switch_struct: dict) -> None:
         try:
             inserted_doc = self._switch_collection.insert_one(switch_struct)
@@ -102,7 +135,8 @@ class NetworkDbOperator:
         return {collection['name']: collection["_id"] for collection in self._switch_collection.find()}
 
     def drop_database(self):
-        self._MongoConnector.drop_database('topology')
+        if self._MongoConnector is not None and "switches" in self._network_state_db.list_collection_names():
+            self._MongoConnector.drop_database("topology")
 
     def dump_network_db(self) -> dict:
